Detect masses for multiple spectra

This function takes multiple spectra imported with batch_import_mzxml() function. Then, it parses the argument to detect_mass() function.

Usage

batch_detect_mass(batch_spect, normalize = TRUE, min_int = 1)

Arguments

batch_spect

a list of MS2 scans imported with batch_import_mzxml() function.

normalize

a boolean indicating if the MS2 spectra is normalized by the base peak before proceeding to filter out low intensity signals (normalize = TRUE), if normalize = FALSE the user has to provide the minimum ion count.

min_int

a integer with the minimum ion intensity

Details

Briefly, similarly to MZmine, detect mass filters out low intensity signals that can be attributed to background noise. Here, you can opt to detect masses by the raw ion intensity or normalize the spectra by the peak base ion and detect masses by intensity percentage from 1% (by default) to 100%.

Examples

# Select the csv file name and path
batch_file <- system.file("extdata", "batch_read.csv",
                          package = "MS2extract")
# Read the data frame
batch_data <- read.csv(batch_file)

# File paths for Procyanidin A2 and Rutin
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")
Rutin_file <- system.file("extdata",
                       "Rutin_neg_20eV.mzXML",
                        package = "MS2extract")

# Add file path - User should specified the file path -
batch_data$File <- c(ProcA2_file, Rutin_file)

# Checking batch_data data frame
batch_data
#>             Name   Formula Ionization_mode min_rt max_rt
#> 1 Procyanidin A2 C30H24O12        Negative    163    180
#> 2          Rutin C27H30O16        Negative    162    171
#>                                                                              File
#> 1 /home/runner/work/_temp/Library/MS2extract/extdata/ProcyanidinA2_neg_20eV.mzXML
#> 2         /home/runner/work/_temp/Library/MS2extract/extdata/Rutin_neg_20eV.mzXML

# Using batch import to import multiple compounds
batch_compounds <- batch_import_mzxml(batch_data)
#> Reading MS2 data from ProcyanidinA2_neg_20eV.mzXML
#> Processing...
#> Reading MS2 data from Rutin_neg_20eV.mzXML
#> Processing...
# Checking dimension by compound
# Procyanidin A2: 24249 ions
# Rutin: 22096 ions
purrr::map(batch_compounds, dim)
#> $`Procyanidin A2`
#> [1] 24249     4
#> 
#> $Rutin
#> [1] 22096     4
#> 

# Extracting MS2 spec
batch_extracted <- batch_extract_MS2(batch_compounds,
                                     verbose = TRUE,
                                     out_list = FALSE)
#> Error in purrr::map(batch_spect, .f = extract_MS2, verbose = verbose,     oiut_list = out_list): ℹ In index: 1.
#> ℹ With name: Procyanidin A2.
#> Caused by error in `.f()`:
#> ! unused argument (oiut_list = FALSE)

#Batch detect mass
batch_mass_detected <- batch_detect_mass(batch_extracted, # Compound list
                                         normalize = TRUE, # Normalize
                                         min_int = 1) # Minimum intensity
#> Error in vctrs_vec_compat(.x, .purrr_user_env): object 'batch_extracted' not found

# Checking dimension by compound
# Procyanidin A2: 38 ions
# Rutin: 4 ions
purrr::map(batch_mass_detected, dim)
#> Error in vctrs_vec_compat(.x, .purrr_user_env): object 'batch_mass_detected' not found