Check spec_metadata for GNPS .mgf format before exporting the MS/MS library

Internal function

Usage

check_specdt_mgf_gnps(spec_metadata = NULL)

Arguments

spec_metadata

a data frame containing the values to be including in the resulting .mgffile. In this case, this is the minimum mandatory information to be included.

Details

This function aims to create a unique KEY to sort and iterate the compound table in order to match the keys created when MS/MS was imported in sort_compound_table().

The three main steps are:

Create a unique row key: Compound name + Polarity + CE

Check if there are any duplicated values

Sort the compound table by Name and CE value