Compare two spectra based on cosine score

Wrap function to calculate the cosine score between two spectra. This function select the *mz* and *intensity* columns before parsing the data frames to OrgMassSpecR::SpectrumSimilarity().

Usage

compare_spectra(spec1, spec2, output.list = T, ...)

Arguments

spec1

a data frame containing the spectra info.

spec2

a data fram contaning the spectra info.

output.list

a boolean if the the output is returned as a list

...

arguments parsed to OrgMassSpecR::SpectrumSimilarity().

Examples

# Importing the Spectrum of Procyanidin A2 in negative ionzation mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

# Importing the MS2 of Procyanidin A2 deconvoluted by PCDL (Agilent)
ProcA2_pcdl_fl <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV_PCDL.csv",
                        package = "MS2extract")

# Reading the Procyanidin A2 spectra
ProcA2_raw <- import_mzxml(ProcA2_file)
#> Reading MS2 data from ProcyanidinA2_neg_20eV.mzXML
#> Processing...
# Normalizing total ion counts (Already normalized)
ProcA2_norm <- normalize_spec(ProcA2_norm)
#> Error in dplyr::mutate(spec, intensity = round(intensity/max(intensity),     2) * 100): object 'ProcA2_norm' not found

# Reading the MS2 spectra of Procynidin A2 by PCDL
ProcA2_PCDL <- read.csv(ProcA2_pcdl_fl)
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former
#> Error in read.table(file = file, header = header, sep = sep, quote = quote,     dec = dec, fill = fill, comment.char = comment.char, ...): no lines available in input

compare_spectra(ProcA2_norm, ProcA2_PCDL)
#> Error in dplyr::select(spec1, mz, intensity): object 'ProcA2_norm' not found