Export MS/ms spectra to a .msp file

Internal function

Usage

write_msp_base(spec = NULL, spec_metadata = NULL)

Arguments

spec

a data frame containing the extracted MS/MS spectra, the following columns are required:

mz_precursor

rt

mz

intensity

spec_metadata

a data frame containing the values to be including in the resulting .msp file. The following column are required as vital information for the .msp output.

NAME

PRECURSORTYPE

FORMULA

RETENTIONTIME

IONMODE

The following fields are also included in the resulting msp files, but are not required to be present in the metadata table. If the column does not exist in the column or the value is missing, it will export a blank field.

INCHIKEY

SMILES

CCS

COLLISIONENERGY

INSTRUMENTTYPE

Details

This functions takes an extracted MS/MS spectra and writes it to a .msp file format. This function incorporates the extracted MS/ms spectra along with metadata for the compound.