Imports mzXML/mzML files with MS2 scans

This function reads .mzXML and .mzML files containing MS2 data using masstools::read_mzxml() and import the data as a list. Each element in a list list represents one scan. Each element in the list of scans contains two sublists that contain 1) the scan information and 2) the spectra per scan.

Usage

import_mzxml(file = NULL, met_metadata = NULL, ppm = 10, ...)

Arguments

file

file name of the mzXML file

met_metadata

a data frame with the following columns. Required:

Formula

A character string specifying the metabolite formula

Ionization_mode

The ionization mode employed in data collection. It can be only Positive or Negative

These two columns are mandatory since the formula is used by the Rdisop package to calculate the exact mass and the ionization mode determine whether the the mass of the a proton is added or subtracted to calculate the charged mass.

Additionally, you can provide the minimum and maximum retention times to look for a peak only within a given area by including the folowing columns.

min_rt

a double with the minimum retention time to keep

max_rt

a double with the minimum retention time to keep

These two columns are highly recommended to be included to narrow down the search window and ensure the peak you want is selected. This is especially important if you have multiple peaks in an EIC.

ppm

the mass error in ppm. 10 ppm is the default value.

...

extra arguments passed to masstools::read_mzxml()

Value

data.frame in a tidy format for MS2 spectra in a tidy format.

mz

ion m/z value

intensity

ion intensity count

precursor_mz

precursor ion

rt

retention time (in seconds)

Examples

# \donttest{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")
# Compound metadata without ROI information
ProcA2_data <- data.frame(Formula = "C30H24O12",Ionization_mode = "Negative")
ProcA2_raw <- import_mzxml(ProcA2_file, met_metadata = ProcA2_data)
#> Reading MS2 data from ProcyanidinA2_neg_20eV.mzXML
#> Processing...

# 26731 ions detected in total
dim(ProcA2_raw)
#> [1] 26731     4

# Region of interest table (rt in seconds)
ProcA2_data <- data.frame(Formula = "C30H24O12",Ionization_mode = "Negative",
                     min_rt = 163, max_rt = 180)
ProcA2_roi <- import_mzxml(ProcA2_file, met_metadata = ProcA2_data)
#> Reading MS2 data from ProcyanidinA2_neg_20eV.mzXML
#> Processing...

# 24249 ions detected in ROI
dim(ProcA2_roi)
#> [1] 24249     4

# }